Add support for additional root CA certs
We're working on getting this pipeline working internally. The first
thing we hit is that Red Hat has its own internal root CA.

We'll need to handle this both for cosa and the Jenkins agent which
wants to fetch `remoting.jar` from the controller as soon as it starts.

This patch handles just the latter for now. We do this by optionally
supporting building our own Jenkins agent with the certificates in it.

For cosa, I think we will be able to work around this by directly
pointing e.g. git and libdnf at the certificate.